report impl

+++ uikit/src/main/java/ru/niku/uikit/PayLoadModel.kt
@@ -3,7 +3,7 @@ package ru.niku.uikit
 class PayLoadModel (
     var id: Int = 0,
     var name: String = "",
-    var amount: Int = 0,
+    var amount: Double = 0.0,
     var category: String = "",
     //var time: Timestamp
 ): BaseValueModel()
